Reload BoringdroidSystemUI after tuner changed
We listen the setting secure key that tuner service changes, and use
PluginManager action to notify the PluginManager to reload this plugin
instance again to reload hook resources of the BoringdroidSystemUI,
to fix the problem that the BoringdroidSystemUI disappears after tuner
changed, because of NavigationBarInflaterView will clear views after
tuner changed.

Also this CL reads listening tuner keys from system properties, so the
ROM developer can add their custom tuner keys from device config with
system properties to support them without changing the source code of
the BoringdroidSystemUI or frameworks/base.

It works fine with simple operations. If we find other problems of this
method, we will try to fix it.

Signed-off-by: utzcoz <utzcoz@outlook.com>