Remove view immediately when onDestroy
In Android 11, the onDestroy is called on non-main thread, so we should
use post to remove views. But in Android 9, it is called on main thread,
we should remove views immediately, otherwise it will kept on
navigation bar group, and the final app state layout will be inserted to
the wrong position.

Signed-off-by: utzcoz <utzcoz@outlook.com>